pyproject.toml
[tool.poetry]
name = "quark-backend"
version = "0.1.0"
description = "Quark P2P Lending Platform API"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.21"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
pydantic = {extras = ["email"], version = "^1.10.13"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
python-dateutil = "^2.9.0.post0"
passlib = "^1.7.4"
bcrypt = "4.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"